Add ability to export input as a Stateless Programmable Switch
When input mode of a switch is set to "Detached", the corresponding
input will be exported as a Stateless Programmable Switch.
It will no longer affect output in any way.

https://github.com/mongoose-os-apps/shelly-homekit/issues/9

+++ src/shelly_input.cpp
@@ -126,7 +126,7 @@ void InputPin::DetectReset(double now, bool cur_state) {
 
 void InputPin::HandleGPIOInt() {
   bool cur_state = GetState();
-  LOG(LL_DEBUG, ("Input %d: %s, st %d", id(), OnOff(cur_state), state_));
+  LOG(LL_DEBUG, ("Input %d: %s, st %d", id(), OnOff(cur_state), (int) state_));
   CallHandlers(Event::kChange, cur_state);
   double now = mgos_uptime();
   DetectReset(now, cur_state);
@@ -168,7 +168,7 @@ void InputPin::HandleTimer() {
   timer_id_ = MGOS_INVALID_TIMER_ID;
   timer_cnt_++;
   bool cur_state = GetState();
-  LOG(LL_DEBUG, ("Input %d: timer, st %d", id(), state_));
+  LOG(LL_DEBUG, ("Input %d: timer, st %d", id(), (int) state_));
   switch (state_) {
     case State::kIdle:
       break;
